# ai-services/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API server for managing AI Services catalog, applications, and authentication",
        "title": "AI Services Catalog API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/project-ai-services/ai-services",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/applications/{id}/datasources": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Links one or more datasource connectors to each eligible (Digitize) service in a running application. Each datasource is processed independently. Returns 422 when no eligible running service with an API endpoint is found.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Datasources"
                ],
                "summary": "Connect datasources to application",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Application ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "List of datasource IDs to connect",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_project-ai-services_ai-services_internal_pkg_catalog_apiserver_models.ConnectDatasourcesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datasources connected",
                        "schema": {
                            "$ref": "#/definitions/github_com_project-ai-services_ai-services_internal_pkg_catalog_apiserver_models.ConnectDatasourcesResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Application or datasource not found",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "No eligible running service found",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Removes one or more datasource connectors from each eligible (Digitize) service in a running application and deletes the service_dependency records. Each datasource is processed independently; failures are surfaced per item.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Datasources"
                ],
                "summary": "Disconnect datasources from application",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Application ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "List of datasource IDs to disconnect",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_project-ai-services_ai-services_internal_pkg_catalog_apiserver_models.DisconnectDatasourcesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datasources disconnected",
                        "schema": {
                            "$ref": "#/definitions/github_com_project-ai-services_ai-services_internal_pkg_catalog_apiserver_models.DisconnectDatasourcesResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Application not found",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "No eligible running service found",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/datasources": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a paginated list of datasource connectors with optional filters by status and provider.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Datasources"
                ],
                "summary": "List datasource connectors",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number (1-indexed)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Number of items per page (max: 100)",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status: 'connected' or 'offline'",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by provider ID (e.g. 'object_storage', 'file_system')",
                        "name": "provider",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_project-ai-services_ai-services_internal_pkg_catalog_apiserver_models.DatasourceListResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid query parameters",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Validates the request, tests the connection, encrypts credentials, and persists a new datasource connector. Returns 422 if the connection test fails.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Datasources"
                ],
                "summary": "Create datasource connector",
                "parameters": [
                    {
                        "description": "Datasource creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_project-ai-services_ai-services_internal_pkg_catalog_apiserver_models.CreateDatasourceRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Datasource created",
                        "schema": {
                            "$ref": "#/definitions/github_com_project-ai-services_ai-services_internal_pkg_catalog_apiserver_models.CreateDatasourceResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or validation errors",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Provider not found in catalog",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Datasource name already exists",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Connection test failed",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/datasources/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the full details of a datasource connector including non-sensitive metadata and the list of connected services enriched with live sync state from each service's Digitize pod.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Datasources"
                ],
                "summary": "Get a single datasource connector",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Datasource UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datasource detail",
                        "schema": {
                            "$ref": "#/definitions/github_com_project-ai-services_ai-services_internal_pkg_catalog_apiserver_models.GetDatasourceResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid UUID format",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Datasource not found",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a datasource connector by ID. Returns 409 Conflict if the connector is still linked to one or more applications.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Datasources"
                ],
                "summary": "Delete datasource connector",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Datasource connector ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Datasource deleted"
                    },
                    "400": {
                        "description": "Invalid connector ID format",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Datasource not found",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Datasource is connected to one or more applications",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_project-ai-services_ai-services_internal_pkg_catalog_apiserver_models.ConnectDatasourcesRequest": {
            "type": "object",
            "required": [
                "datasource_ids"
            ],
            "properties": {
                "datasource_ids": {
                    "description": "DatasourceIDs is the list of datasource connector UUIDs to connect.",
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "github_com_project-ai-services_ai-services_internal_pkg_catalog_apiserver_models.ConnectDatasourcesResponse": {
            "type": "object",
            "properties": {
                "application_id": {
                    "type": "string"
                },
                "connections": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_project-ai-services_ai-services_internal_pkg_catalog_apiserver_models.DatasourceConnectionItem"
                    }
                }
            }
        },
...
        "github_com_project-ai-services_ai-services_internal_pkg_catalog_apiserver_models.CreateDatasourceResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "github_com_project-ai-services_ai-services_internal_pkg_catalog_apiserver_models.DatasourceConnectionItem": {
            "type": "object",
            "properties": {
                "connector_id": {
                    "type": "string"
                },
                "datasource_id": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
        "github_com_project-ai-services_ai-services_internal_pkg_catalog_apiserver_models.DatasourceListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_project-ai-services_ai-services_internal_pkg_catalog_apiserver_models.DatasourceResponse"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/github_com_project-ai-services_ai-services_internal_pkg_catalog_types.PaginationMetadata"
                }
            }
        },
...
        "github_com_project-ai-services_ai-services_internal_pkg_catalog_apiserver_models.DisconnectDatasourcesRequest": {
            "type": "object",
            "required": [
                "datasource_ids"
            ],
            "properties": {
                "datasource_ids": {
                    "description": "DatasourceIDs is the list of datasource connector UUIDs to disconnect.",
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "github_com_project-ai-services_ai-services_internal_pkg_catalog_apiserver_models.DisconnectDatasourcesResponse": {
            "type": "object",
            "properties": {
                "application_id": {
                    "type": "string"
                },
                "disconnections": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_project-ai-services_ai-services_internal_pkg_catalog_apiserver_models.DatasourceDisconnectionItem"
                    }
                }
            }
        },
...
        "internal_pkg_catalog_apiserver_handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}